feat(setup): frontend and backend frameworks
Went with rocket for the backend, this seems similar to flask so might
be easier to get the hang of. Requires nightly so make sure to `rustup
set overide nightly`.

Using yew and wasm-pack to build the frontend. This hopefully means we
won't need to write any js. Might be more effort than it's worth, we can
always switch to something else later on.