feat(player-view): inital commit on player_view component
Not really much in the component that at the moment. This is more an
example of how comomponents _might_ fit nicely together. Really enjoying
the way that the front and backends will be able to share structures

TODO: There is currently no way to get a single players hand from the backend
TODO: work out how to call the backend with yew